firmware: properly initialize Save RAM area
The need to save is determined based on a CRC change of the contents of the save RAM area.
However the first CRC sample was taken shortly after running the console. Thus, if software writes to save RAM very early we won't notice.

Save RAM is now zeroed before attempting to load an existing SRM file and the initial checksum is taken BEFORE running the game.