in-game hooks: more time efficient Auto Joypad Read snooping
This is a bit of a fire-and-forget strategy. After NMI/IRQ read the AJR
result register as fast as possible to catch the readout from last
frame.
In some occasions AJR has already JUST started (its starting point is
variable) in which case the readout is #$0000. For IRQ hooks this might
lead to unusable readouts depending on the time of the IRQ.
Also don't compare controller button mappings when no buttons are
pressed.