Enhance menu visually (disabled entries, change hooks)
Menu entries can now be printed in grey when some condition is (not)
met. A function call is used to determine the status with the carry flag
as a boolean return value. Disabled menu entries can still be changed
for when they become active again.
Also a change hook function can now be used in the menu structure. It is
called whenever the value is changed _during setting_.