Build: new version handling
Version is now generally pulled from the RELEASE_VERSION variable in
`src/VERSION`.
`VERSION=...` can be passed to make to override RELEASE_VERSION. The
special value `VERSION=*` will use the current timestamp as the version
(`YYmmddHHMMSS`). If neither VERSION nor RELEASE_VERSION are specified,
`SNAPSHOT` is used as the version.