Workaround S-CPU A DMA silicon bug
The SNES's CPU revision A occasionally does not carry out DMA transfers when a HDMA channel is configured to use $2100 as a B-Bus target, i.e. $00 must not be written to any DMA channel's B-Bus address register at any time.
The fix sets $21ff (unmapped) as a target address instead and configures the DMA channel to write two registers, thereby wrapping around to $2100 on the second write.